criacao de butao edit e modal

+++ BackOffice/userModule/__init__.py
@@ -14,26 +14,33 @@ def toUsersPage():
         get_Users = db.session.query(User).all()
         return render_template("users.html", title="Login", get_Users = get_Users)
     elif request.method == 'POST':
-        id = request.form.get("id")
+        if request.form.get("action")=="Apagar":
+            id = request.form.get("id")
         
-        specify_User = db.session.query(User).filter(User.id==id).first()
-        specify_Ride = db.session.query(Ride).filter(Ride.user_id==specify_User.id).first()
-        specify_Vehicle = db.session.query(Vehicle).filter(Vehicle.owner_id==specify_User.id).first()
-        specify_Reservation = db.session.query(Reservation).filter(Reservation.passenger_id==specify_User.id).first()
+            specify_User = db.session.query(User).filter(User.id==id).first()
+            specify_Ride = db.session.query(Ride).filter(Ride.user_id==specify_User.id).first()
+            specify_Vehicle = db.session.query(Vehicle).filter(Vehicle.owner_id==specify_User.id).first()
+            specify_Reservation = db.session.query(Reservation).filter(Reservation.passenger_id==specify_User.id).first()
 
-        if(specify_Ride != None):
-            db.session.delete(specify_Ride)
+            if(specify_Ride != None):
+                db.session.delete(specify_Ride)
 
-        if(specify_Vehicle != None):
-            db.session.delete(specify_Vehicle)
-            
-        if(specify_Reservation != None):
-            db.session.delete(specify_Reservation)
+            if(specify_Vehicle != None):
+                db.session.delete(specify_Vehicle)
 
-        db.session.delete(specify_User)
-        db.session.commit()
-        
-    return  jsonify(message="Apagado",status=201)
+            if(specify_Reservation != None):
+                db.session.delete(specify_Reservation)
+
+            db.session.delete(specify_User)
+            db.session.commit()
+
+            return  jsonify(message="Apagado",status=201)
+
+        if request.form.get("action")=="Editar":
+            idUser = request.form.get("id")
+            userEdit = db.session.query(User).filter(User.id==idUser).first()
+            print(userEdit)
+            return jsonify()
